acasb-analysis/requirements-experimental.txt
torch>=2.2,<2.7
torchvision>=0.17,<0.22
Pillow>=10,<12
